+++ libraries/stdlib/clr/kotlin/collections/CollectionsClr.kt
@@ -19,6 +19,7 @@ import kotlin.internal.InlineOnly
  *
  * @sample samples.collections.Collections.Lists.singletonReadOnlyList
  */
+@kotlin.clr.ClrCollectionFactory("list")
 public actual fun <T> listOf(element: T): List<T> {
     val list = ArrayList<T>(1)
     list.add(element)